+++ pubmodules/whatsapp/listener.js
@@ -27,6 +27,12 @@ class Listener {
 
         let baseFileUrl = process.env.BASE_FILE_URL || apiUrl || "http://localhost:3000"
 
+        let job_topic = process.env.JOB_TOPIC_EXCHANGE;
+        winston.debug("Whatsapp job topic " + job_topic);
+
+        let amqp_manager_url = process.env.AMQP_MANAGER_URL;
+        winston.debug("amqp_manager_url " + amqp_manager_url);
+
         let log = process.env.WHATSAPP_LOG || false
         winston.debug("Whatsapp log: "+ log);
 
@@ -40,6 +46,8 @@ class Listener {
             REDIS_HOST: host,
             REDIS_PORT: port,
             REDIS_PASSWORD: password,
+            AMQP_MANAGER_URL: amqp_manager_url,
+            JOB_TOPIC_EXCHANGE: job_topic,
             log: log
         }, (err) => {
             if (!err) {