+++ CHANGELOG.md
@@ -5,6 +5,22 @@
 🚀        IN PRODUCTION                        🚀
 (https://www.npmjs.com/package/@tiledesk/tiledesk-server/v/2.3.77) 
 
+# 2.4.83
+- Improved whatsapp log services
+- Updated whatsapp-connector to 0.1.57
+- Updated whatsapp-jobworker to 0.0.7
+
+# 2.4.82
+- Added whatsapp log services
+
+# 2.4.81
+- update whatsapp-connector to 0.1.56
+
+# 2.4.79
+- update whatsapp-connector to 0.1.55
+- update whatsapp-jobworker to 0.0.4
+- added support for whatsapp broadcast queue
+
 # 2.4.78
 - update whatsapp-connector to 0.1.53
 - update messenger-connector to 0.1.13